+++ src/App.js
@@ -2,7 +2,7 @@ import React, { useEffect } from 'react';
 import { ThemeProvider } from 'styled-components';
 import { BrowserRouter, Route, Switch, Redirect } from 'react-router-dom';
 import { WalletContextProvider } from '@provenanceio/wallet-lib';
-import { useApp, useAssets } from 'redux/hooks';
+import { useApp, useAssets, useMediaQuery } from 'redux/hooks';
 import { Navigation, Footer, SpriteSheet, BaseStyle } from 'Components';
 import { GlobalStyle, Themes } from 'theme';
 import { isProd } from 'consts';
@@ -25,8 +25,10 @@ import {
 
 const App = () => {
   const { theme, walletUrl } = useApp();
+  // Support OS theme settings if the user hasn't clicked the switch at the top
+  const { matches: dark } = useMediaQuery('(prefers-color-scheme: dark)');
+  const activeTheme = Themes[theme] ? Themes[theme] : dark ? Themes.night : Themes.default;
   const { assetMetadata, assetMetadataLoading, getAssetMetadata } = useAssets();
-  const activeTheme = Themes[theme] || Themes.default;
 
   useEffect(() => {
     if (isEmpty(assetMetadata) && !assetMetadataLoading) {